# src/FiapCloudGames.Domain/DTOs/AuthResponseDto.cs
namespace FiapCloudGames.Domain.DTOs
{
    public class AuthResponseDto
    {
        public string Token { get; set; } = string.Empty;
        public string Email { get; set; } = string.Empty;
        public string Name { get; set; } = string.Empty;
        public int UserId { get; set; }
    }
}