# src/FiapCloudGames.Domain/DTOs/LoginDto.cs
namespace FiapCloudGames.Domain.DTOs
{
    using System.ComponentModel.DataAnnotations;

    public class LoginDto
    {
        [Required(ErrorMessage = "O e-mail é obrigatório.")]
        [EmailAddress(ErrorMessage = "Formato de e-mail inválido.")]
        [StringLength(255, ErrorMessage = "O e-mail deve ter no máximo 255 caracteres.")]
        public required string Email { get; set; }

        [Required(ErrorMessage = "A senha é obrigatória.")]
        [StringLength(128, MinimumLength = 1, ErrorMessage = "A senha é obrigatória.")]
        public required string Password { get; set; }
    }
}